Improve parsing of arguments and options
Improving parsing of arguments and options by using the standard library
"argparse". While I was at it, I made the 'path' to the root of the PDFs
files' paths optional, as it can be guessed from the path of the BibDesk
BibTeX file. Furthermore, the tool now supports (and is actually the
default) read from standard input and the output file can be configured
through an option (useful if the environment does not support output
redirection).

The new usage string (generated automatically by argparse) is the
following:
bibdesk2zotero [--help] [--out OUT] [--path PATH] [database]